items: refactor items modules
This separates out related item functions into their own modules to
reduce the scope of the common module.
