+++ src/libtrx/include/libtrx/game/items.h
@@ -1,8 +1,9 @@
 #pragma once
 
+#include "items/anim.h"
+#include "items/col.h"
 #include "items/common.h"
 #include "items/const.h"
 #include "items/enum.h"
-#include "items/misc.h"
 #include "items/types.h"
 #include "items/utils.h"
