--- src/libtrx/include/libtrx/game/items/misc.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "../anims/types.h"
-#include "types.h"
-
-ANIM_FRAME *Item_GetBestFrame(const ITEM *item);
