+++ src/libtrx/include/libtrx/game/items/utils.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "./types.h"
+
 #define ITEM_ADJUST_ROT(source, target, rot)                                   \
     do {                                                                       \
         if ((int16_t)(target - source) > rot) {                                \
@@ -10,3 +12,12 @@
             source = target;                                                   \
         }                                                                      \
     } while (0)
+
+void Item_TakeDamage(ITEM *item, int16_t damage, bool hit_status);
+
+// Mesh_bits: which meshes to affect.
+// Damage:
+// * Positive values - deal damage, enable body part explosions.
+// * Negative values - deal damage, disable body part explosions.
+// * Zero - don't deal any damage, disable body part explosions.
+int32_t Item_Explode(int16_t item_num, int32_t mesh_bits, int16_t damage);
