+++ src/libtrx/meson.build
@@ -159,7 +159,10 @@ sources = [
   'game/interpolation.c',
   'game/inventory.c',
   'game/inventory_ring/priv.c',
-  'game/items.c',
+  'game/items/anim.c',
+  'game/items/col.c',
+  'game/items/common.c',
+  'game/items/utils.c',
   'game/lara/common.c',
   'game/lara/misc.c',
   'game/level/common.c',
